refactor: 调整脚本路径
