+++ src/BootstrapBlazor.Server/Components/Samples/Icons/FAIconList.razor
@@ -1,5 +1,4 @@
-﻿@namespace BootstrapBlazor.Components
-@inherits BootstrapModuleComponentBase
+﻿@inherits WebSiteModuleComponentBase
 
 <div @attributes="@AdditionalAttributes" class="@ClassString" data-bb-title="@CopiedTooltipText" id="@Id">
     <div class="icons-body" data-bs-spy="scroll" data-bs-target=".icons-nav" data-bs-smooth-scroll="true" tabindex="0">
